stm32g0-sys driver
This makes things small enough that I can fit hiffy onto the g031.
