--- drv/stm32g0-gpio-api/build.rs
@@ -1,11 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at https://mozilla.org/MPL/2.0/.
-
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-    idol::client::build_client_stub(
-        "../../idl/stm32g0-gpio.idol",
-        "client_stub.rs",
-    )?;
-    Ok(())
-}
