--- drv/stm32g0-gpio/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "drv-stm32g0-gpio"
-version = "0.1.0"
-edition = "2018"
-
-[dependencies]
-userlib = {path = "../../sys/userlib"}
-zerocopy = "0.6.1"
-byteorder = { version = "1.3.4", default-features = false }
-num-traits = { version = "0.2.12", default-features = false }
-drv-stm32g0-rcc-api = {path = "../stm32g0-rcc-api", default-features = false }
-stm32g0 = { git = "https://github.com/oxidecomputer/stm32-rs-nightlies", branch = "stm32g0b1-initial-support", default-features = false }
-cortex-m = { version = "0.7", features = ["inline-asm"] }
-unwrap-lite = {path = "../../lib/unwrap-lite"}
-
-[features]
-g031 = ["stm32g0/stm32g031", "drv-stm32g0-rcc-api/g031"]
-g070 = ["stm32g0/stm32g070", "drv-stm32g0-rcc-api/g070"]
-g0b1 = ["stm32g0/stm32g0b1", "drv-stm32g0-rcc-api/g0b1"]
-
-# This section is here to discourage RLS/rust-analyzer from doing test builds,
-# since test builds don't work for cross compilation.
-[[bin]]
-name = "drv-stm32g0-gpio"
-test = false
-bench = false
