--- drv/stm32g0-rcc-api/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "drv-stm32g0-rcc-api"
-version = "0.1.0"
-edition = "2018"
-
-[dependencies]
-userlib = {path = "../../sys/userlib"}
-zerocopy = "0.6.1"
-byteorder = {version = "1.3", default-features = false}
-num-traits = {version = "0.2", default-features = false}
-unwrap-lite = {path = "../../lib/unwrap-lite"}
-
-[build-dependencies]
-idol = {git = "https://github.com/oxidecomputer/idolatry.git"}
-
-[features]
-g031 = []
-g070 = []
-g0b1 = []
-
-# This section is here to discourage RLS/rust-analyzer from doing test builds,
-# since test builds don't work for cross compilation.
-[lib]
-test = false
-bench = false
