+++ drv/stm32g0-usart/Cargo.toml
@@ -7,16 +7,15 @@ edition = "2018"
 userlib = {path = "../../sys/userlib"}
 zerocopy = "0.6.1"
 num-traits = { version = "0.2.12", default-features = false }
-drv-stm32g0-gpio-api = {path = "../stm32g0-gpio-api"}
-drv-stm32g0-rcc-api = {path = "../stm32g0-rcc-api", default-features = false}
+drv-stm32g0-sys-api = {path = "../stm32g0-sys-api", default-features = false}
 stm32g0 = { git = "https://github.com/oxidecomputer/stm32-rs-nightlies", branch = "stm32g0b1-initial-support", default-features = false }
 cortex-m = { version = "0.7", features = ["inline-asm"] }
 cortex-m-semihosting = { version = "0.3.7", features = ["inline-asm"] }
 
 [features]
-g031 = ["stm32g0/stm32g031", "drv-stm32g0-rcc-api/g031"]
-g070 = ["stm32g0/stm32g070", "drv-stm32g0-rcc-api/g070"]
-g0b1 = ["stm32g0/stm32g0b1", "drv-stm32g0-rcc-api/g0b1"]
+g031 = ["stm32g0/stm32g031", "drv-stm32g0-sys-api/g031"]
+g070 = ["stm32g0/stm32g070", "drv-stm32g0-sys-api/g070"]
+g0b1 = ["stm32g0/stm32g0b1", "drv-stm32g0-sys-api/g0b1"]
 semihosting = ["userlib/log-semihosting"]
 
 # This section is here to discourage RLS/rust-analyzer from doing test builds,
