+++ idl/stm32g0-sys.idol
@@ -1,0 +1,103 @@
+// STM32G0 "system" IPC API
+
+Interface(
+    name: "Sys",
+    ops: {
+        "enable_clock_raw": (
+            args: {
+                "peripheral": "u32",
+            },
+            reply: Result(
+                ok: "()",
+                err: CLike("RccError"),
+            ),
+            idempotent: true,
+        ),
+        "disable_clock_raw": (
+            args: {
+                "peripheral": "u32",
+            },
+            reply: Result(
+                ok: "()",
+                err: CLike("RccError"),
+            ),
+            idempotent: true,
+        ),
+        "enter_reset_raw": (
+            args: {
+                "peripheral": "u32",
+            },
+            reply: Result(
+                ok: "()",
+                err: CLike("RccError"),
+            ),
+            idempotent: true,
+        ),
+        "leave_reset_raw": (
+            args: {
+                "peripheral": "u32",
+            },
+            reply: Result(
+                ok: "()",
+                err: CLike("RccError"),
+            ),
+            idempotent: true,
+        ),
+        "gpio_configure_raw": (
+            args: {
+                "port": (
+                    type: "Port",
+                    recv: FromPrimitive("u8"),
+                ),
+                "pins": "u16",
+                "packed_attributes": "u16",
+            },
+            reply: Result(
+                ok: "()",
+                err: CLike("GpioError"),
+            ),
+            idempotent: true,
+        ),
+        "gpio_set_reset": (
+            args: {
+                "port": (
+                    type: "Port",
+                    recv: FromPrimitive("u8"),
+                ),
+                "set_pins": "u16",
+                "reset_pins": "u16",
+            },
+            reply: Result(
+                ok: "()",
+                err: CLike("GpioError"),
+            ),
+            idempotent: true,
+        ),
+        "gpio_read_input": (
+            args: {
+                "port": (
+                    type: "Port",
+                    recv: FromPrimitive("u8"),
+                ),
+            },
+            reply: Result(
+                ok: "u16",
+                err: CLike("GpioError"),
+            ),
+            idempotent: true,
+        ),
+        "gpio_toggle": (
+            args: {
+                "port": (
+                    type: "Port",
+                    recv: FromPrimitive("u8"),
+                ),
+                "pins": "u16",
+            },
+            reply: Result(
+                ok: "()",
+                err: CLike("GpioError"),
+            ),
+        ),
+    },
+)
