+++ task/hiffy/Cargo.toml
@@ -14,7 +14,7 @@ drv-stm32h7-i2c = {path = "../../drv/stm32h7-i2c", default-features = false, opt
 drv-stm32h7-rcc-api = {path = "../../drv/stm32h7-rcc-api", default-features = false, optional = true }
 drv-lpc55-gpio-api = {path = "../../drv/lpc55-gpio-api", optional = true}
 drv-gimlet-hf-api = {path = "../../drv/gimlet-hf-api", optional = true}
-drv-stm32g0-gpio-api = {path = "../../drv/stm32g0-gpio-api", default-features = false, optional = true}
+drv-stm32g0-sys-api = {path = "../../drv/stm32g0-sys-api", default-features = false, optional = true}
 cortex-m = { version = "0.7", features = ["inline-asm"] }
 cfg-if = "0.1.10"
 zerocopy = "0.6.1"
@@ -38,7 +38,7 @@ gpio = []
 spi = []
 stm32h7 = ["drv-stm32h7-gpio-api", "userlib/panic-messages"]
 lpc55 = ["drv-lpc55-gpio-api", "userlib/panic-messages"]
-stm32g0 = ["drv-stm32g0-gpio-api"]
+stm32g0 = ["drv-stm32g0-sys-api"]
 qspi = ["drv-gimlet-hf-api"]
 h743 = ["drv-stm32h7-rcc-api/h743", "drv-stm32h7-i2c/h743", "build-i2c/h743"]
 h753 = ["drv-stm32h7-rcc-api/h753", "drv-stm32h7-i2c/h753", "build-i2c/h753"]
