+++ task/hiffy/src/stm32g0.rs
@@ -3,12 +3,11 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 use hif::*;
-use hubris_num_tasks::Task;
 #[cfg(any(feature = "gpio"))]
 use userlib::*;
 
 #[cfg(feature = "gpio")]
-task_slot!(GPIO, gpio_driver);
+task_slot!(SYS, sys);
 
 pub struct Buffer(u8);
 
@@ -19,41 +18,41 @@ pub struct Buffer(u8);
 pub enum Functions {
     Sleep(u16, u32),
     #[cfg(feature = "gpio")]
-    GpioInput(drv_stm32g0_gpio_api::Port, drv_stm32g0_gpio_api::GpioError),
+    GpioInput(drv_stm32g0_sys_api::Port, drv_stm32g0_sys_api::GpioError),
     #[cfg(feature = "gpio")]
     GpioToggle(
-        (drv_stm32g0_gpio_api::Port, u8),
-        drv_stm32g0_gpio_api::GpioError,
+        (drv_stm32g0_sys_api::Port, u8),
+        drv_stm32g0_sys_api::GpioError,
     ),
     #[cfg(feature = "gpio")]
     GpioSet(
-        (drv_stm32g0_gpio_api::Port, u8),
-        drv_stm32g0_gpio_api::GpioError,
+        (drv_stm32g0_sys_api::Port, u8),
+        drv_stm32g0_sys_api::GpioError,
     ),
     #[cfg(feature = "gpio")]
     GpioReset(
-        (drv_stm32g0_gpio_api::Port, u8),
-        drv_stm32g0_gpio_api::GpioError,
+        (drv_stm32g0_sys_api::Port, u8),
+        drv_stm32g0_sys_api::GpioError,
     ),
     #[cfg(feature = "gpio")]
     GpioConfigure(
         (
-            drv_stm32g0_gpio_api::Port,
+            drv_stm32g0_sys_api::Port,
             u8,
-            drv_stm32g0_gpio_api::Mode,
-            drv_stm32g0_gpio_api::OutputType,
-            drv_stm32g0_gpio_api::Speed,
-            drv_stm32g0_gpio_api::Pull,
-            drv_stm32g0_gpio_api::Alternate,
+            drv_stm32g0_sys_api::Mode,
+            drv_stm32g0_sys_api::OutputType,
+            drv_stm32g0_sys_api::Speed,
+            drv_stm32g0_sys_api::Pull,
+            drv_stm32g0_sys_api::Alternate,
         ),
-        drv_stm32g0_gpio_api::GpioError,
+        drv_stm32g0_sys_api::GpioError,
     ),
 }
 
 #[cfg(feature = "gpio")]
 fn gpio_args(
     stack: &[Option<u32>],
-) -> Result<(drv_stm32g0_gpio_api::Port, u16), Failure> {
+) -> Result<(drv_stm32g0_sys_api::Port, u16), Failure> {
     if stack.len() < 2 {
         return Err(Failure::Fault(Fault::MissingParameters));
     }
@@ -61,7 +60,7 @@ fn gpio_args(
     let fp = stack.len() - 2;
 
     let port = match stack[fp + 0] {
-        Some(port) => match drv_stm32g0_gpio_api::Port::from_u32(port) {
+        Some(port) => match drv_stm32g0_sys_api::Port::from_u32(port) {
             Some(port) => port,
             None => return Err(Failure::Fault(Fault::BadParameter(0))),
         },
@@ -89,8 +88,8 @@ fn gpio_input(
 ) -> Result<usize, Failure> {
     use byteorder::ByteOrder;
 
-    let task = GPIO.get_task_id();
-    let gpio = drv_stm32g0_gpio_api::Gpio::from(task);
+    let task = SYS.get_task_id();
+    let sys = drv_stm32g0_sys_api::Sys::from(task);
 
     if stack.len() < 1 {
         return Err(Failure::Fault(Fault::MissingParameters));
@@ -99,14 +98,14 @@ fn gpio_input(
     let fp = stack.len() - 1;
 
     let port = match stack[fp + 0] {
-        Some(port) => match drv_stm32g0_gpio_api::Port::from_u32(port) {
+        Some(port) => match drv_stm32g0_sys_api::Port::from_u32(port) {
             Some(port) => port,
             None => return Err(Failure::Fault(Fault::BadParameter(0))),
         },
         None => return Err(Failure::Fault(Fault::EmptyParameter(0))),
     };
 
-    match gpio.read_input(port) {
+    match sys.gpio_read_input(port) {
         Ok(input) => {
             byteorder::LittleEndian::write_u16(rval, input);
             Ok(core::mem::size_of::<u16>())
@@ -121,12 +120,12 @@ fn gpio_toggle(
     _data: &[u8],
     _rval: &mut [u8],
 ) -> Result<usize, Failure> {
-    let task = GPIO.get_task_id();
-    let gpio = drv_stm32g0_gpio_api::Gpio::from(task);
+    let task = SYS.get_task_id();
+    let sys = drv_stm32g0_sys_api::Sys::from(task);
 
     let (port, mask) = gpio_args(stack)?;
 
-    match gpio.toggle(port, mask) {
+    match sys.gpio_toggle(port, mask) {
         Ok(_) => Ok(0),
         Err(err) => Err(Failure::FunctionError(err.into())),
     }
@@ -138,12 +137,12 @@ fn gpio_set(
     _data: &[u8],
     _rval: &mut [u8],
 ) -> Result<usize, Failure> {
-    let task = GPIO.get_task_id();
-    let gpio = drv_stm32g0_gpio_api::Gpio::from(task);
+    let task = SYS.get_task_id();
+    let sys = drv_stm32g0_sys_api::Sys::from(task);
 
     let (port, mask) = gpio_args(stack)?;
 
-    match gpio.set_reset(port, mask, 0) {
+    match sys.gpio_set_reset(port, mask, 0) {
         Ok(_) => Ok(0),
         Err(err) => Err(Failure::FunctionError(err.into())),
     }
@@ -155,12 +154,12 @@ fn gpio_reset(
     _data: &[u8],
     _rval: &mut [u8],
 ) -> Result<usize, Failure> {
-    let task = GPIO.get_task_id();
-    let gpio = drv_stm32g0_gpio_api::Gpio::from(task);
+    let task = SYS.get_task_id();
+    let sys = drv_stm32g0_sys_api::Sys::from(task);
 
     let (port, mask) = gpio_args(stack)?;
 
-    match gpio.set_reset(port, 0, mask) {
+    match sys.gpio_set_reset(port, 0, mask) {
         Ok(_) => Ok(0),
         Err(err) => Err(Failure::FunctionError(err.into())),
     }
@@ -172,7 +171,7 @@ fn gpio_configure(
     _data: &[u8],
     _rval: &mut [u8],
 ) -> Result<usize, Failure> {
-    use drv_stm32g0_gpio_api::*;
+    use drv_stm32g0_sys_api::*;
 
     if stack.len() < 7 {
         return Err(Failure::Fault(Fault::MissingParameters));
@@ -221,10 +220,10 @@ fn gpio_configure(
         None => return Err(Failure::Fault(Fault::EmptyParameter(6))),
     };
 
-    let task = GPIO.get_task_id();
-    let gpio = drv_stm32g0_gpio_api::Gpio::from(task);
+    let task = SYS.get_task_id();
+    let sys = drv_stm32g0_sys_api::Sys::from(task);
 
-    match gpio.configure(port, mask, mode, output_type, speed, pull, af) {
+    match sys.gpio_configure(port, mask, mode, output_type, speed, pull, af) {
         Ok(_) => Ok(0),
         Err(err) => Err(Failure::FunctionError(err.into())),
     }
